requirements.txt
pandas
numpy
scikit-learn
lightgbm
torch
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
pyyaml
joblib
requests

